Impelement unicode normalization APIs
Unicode normalization API for ByteString and Text. The API is based on the
bindings to the utf8proc library lifted from the XQilla project.

The normalization API is based on the text-icu package.

The common code is under the Data.Unicode module.
Text and ByteString specific code is in Data.Text.* and Data.ByteString.*
modules respectively.