Add native haskell decompose for experimenting
This is just decompose code which is pretty short and simple to compare
performance and experiment with the data structures to see how best it can
perform compared to icu and utf8proc.

The script to generate haskell data structures from unicode database and the
decompose code adpated from Antonio Nikishaev's prose library.