Remove one "if" condition from the fast path
This is to demonstrate a drastic change in performance by just removing a branch
which was never taken. Performance of the loop improved by around 70%, earlier
it was taking around 4.4 ms and after removing the branch it takes 2.6 ms.

Relevant simpl/stg/cmm/asm are also checked in.

To see how the actual instructions executed in the loop changed do a "git
difftool" on - GHC_PERF_ISSUE/decompose-loop-execution-trace-ghc-7.10.3.asm