Fix perf regression caused by refactor of decomposeChar
Earlier I thought the regression due to refactor of decomposeChar was just a
noise, but usinng performance testing without "--quick" showed a consistent
regression.

The reason was found to be the capture of the "marr" argument in the
closures defined via the "where" clause. Now I remember why I had
written this code in this way originally. In performance sensitive code
we should consider passing arguments on stack instead of capturing them
in the closure.