use lookahead for faster hangul composition
Earlier were determining the type of the hangul syllable as soon as
we saw it.  Determining whether it is a hangul LV or not required a
division operation which is expensive.

Instead we now delay this decision until we see the next character. In
most cases we do not need to determine this, we only need to determine
it if the syllable is LV and it is followed by a T. With this we
need very little division, therefore speeding up hangul composition
significantly.